Move function into util folder
