+++ apps/web/components/Organization/Slice/LatestNewsSlice/DigitalIcelandLatestNewsSlice/DigitalIcelandLatestNewsSlice.tsx
@@ -15,7 +15,7 @@ import { BackgroundImage } from '@island.is/web/components'
 import { FRONTPAGE_NEWS_TAG_ID } from '@island.is/web/constants'
 import { LatestNewsSlice as LatestNewsSliceSchema } from '@island.is/web/graphql/schema'
 import { useLinkResolver } from '@island.is/web/hooks'
-import { shortenText } from '@island.is/web/screens/IcelandicGovernmentInstitutionVacancies/IcelandicGovernmentInstitutionVacanciesList'
+import { shortenText } from '@island.is/web/utils/shortenText'
 
 import * as styles from './DigitalIcelandLatestNewsSlice.css'
